fix: 修改minLineSpacing类型为CGFloat

+++ Sources/FWObjC/Module/View/FWPasscodeView.h
@@ -16,7 +16,7 @@ NS_SWIFT_NAME(PasscodeFlowLayout)
 
 @property (assign, nonatomic) NSInteger itemNum;
 
-@property (assign, nonatomic) NSInteger minLineSpacing;
+@property (assign, nonatomic) CGFloat minLineSpacing;
 
 - (void)updateLineSpacing;
